Started initial theming, added example for populating fragments
Got rid of the action bar, changed default theming of the tabs to be red and white. Yay, RPI colors.
Also, added a simple example of how we can populate the fragments based on arguments. We can extend this to do more than just put a number in a text box.