Transition of location variables to person class
Moved all location related variable to the newly created person class.
This allows us to store a location within the person object, and
therefore access later at the needed time.  Right now, set up so that
location only updates with an explicit call to that person so as to not
unnecessarily update location.  Can change later if needed.  Also
created an example person object in map view to insure it was working.